requirements.txt
django==4.0.6
djangorestframework==3.13.1
PyJWT==2.4.0
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.2.0
docker==5.0.3
docker-compose==1.29.2
flake8==4.0.1